refactor(NODE-3717): Test reorg part 7 (#3093)
