+++ test/integration/auth/scram_sha_256.test.js
@@ -2,10 +2,10 @@
 
 const sinon = require('sinon');
 const { expect } = require('chai');
-const { Connection } = require('../../src/cmap/connection');
-const { ScramSHA256 } = require('../../src/cmap/auth/scram');
-const { setupDatabase, withClient } = require('./shared');
-const { LEGACY_HELLO_COMMAND } = require('../../src/constants');
+const { Connection } = require('../../../src/cmap/connection');
+const { ScramSHA256 } = require('../../../src/cmap/auth/scram');
+const { setupDatabase, withClient } = require('../shared');
+const { LEGACY_HELLO_COMMAND } = require('../../../src/constants');
 
 describe('SCRAM-SHA-256 auth', function () {
   const test = {};
