+++ test/integration/auth/ssl_x509_connect.test.js
@@ -1,10 +1,9 @@
 'use strict';
-var fs = require('fs');
-var f = require('util').format;
-var test = require('./shared').assert;
-var setupDatabase = require('./shared').setupDatabase;
+const fs = require('fs');
+const { format: f } = require('util');
+const { test, setupDatabase } = require('../shared');
 const { expect } = require('chai');
-const { MongoClient } = require('../../src');
+const { MongoClient } = require('../../../src');
 
 describe('SSL (x509)', function () {
   before(function () {
